perf: change app launch flow
Previously, we hide the splash screen as soon as `useCachedResources` finished loading assets.

But because the navigation component is heavy, it causes a blank screen to be shown for ~2sec before the screens rendered.

This commit fixes that by keeping the splash screen visible when assets are being loaded, when `remote-config` is setting its defaults, and when the `PersistGate` is loading the app state. The splash screen only hides in the `useEffect` of `Navigation` component, i..e after it has rendered its children (screens)